Update `permalink` config return value
Changing the return value to a function addresses a weird issue that came up
where templates appeared to be sent to the `compile` function twice. The first
time without an `outputPath` which caused this plugin to break.

Not entirely sure why, but in a test repo, changing the signature here to return
a function solved the problem.